Fix: 수업 리포트 Put API 시, 참여자와 숙제 정보의 변경 여부에 따라 조건 분기 되도록 수정 진행 #173

+++ src/main/java/com/selfrunner/gwalit/domain/lesson/service/LessonService.java
@@ -34,6 +34,7 @@ import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 @Service
@@ -92,17 +93,52 @@ public class LessonService {
         Lesson lesson = lessonRepository.findById(lessonId).orElseThrow(() -> new LessonException(ErrorCode.NOT_EXIST_LESSON));
         memberAndLectureRepository.findMemberAndLectureByMemberAndLectureLectureId(member, lesson.getLecture().getLectureId()).orElseThrow(() -> new MemberException(ErrorCode.UNAUTHORIZED_EXCEPTION));
 
-        // Business Logic
+        // Business Logic: Homework 변경 여부 확인 진행
         lesson.update(putLessonReq);
-        homeworkRepository.deleteHomeworkByLessonId(lessonId);
-        List<Homework> homeworkList = new ArrayList<>();
-        for (Participant participant : putLessonReq.getParticipants()) {
-            List<Homework> tempHomeworkList = putLessonReq.getHomeworks().stream()
-                    .map(homeworkReq -> HomeworkReq.staticToEntity(homeworkReq, participant.getMemberId(), lesson.getLessonId()))
-                    .collect(Collectors.toList());
-            homeworkList.addAll(tempHomeworkList);
+        // 참여자 비교
+        Boolean needUpdate = Boolean.FALSE;
+        if(lesson.getParticipants().size() != putLessonReq.getParticipants().size()) {
+            needUpdate = Boolean.TRUE;
         }
-        homeworkRepository.saveAll(homeworkList);
+        if(lesson.getParticipants().size() == putLessonReq.getParticipants().size()) {
+            for(Participant lessonParticipant : lesson.getParticipants()) {
+                Boolean temp = Boolean.FALSE;
+                for(Participant dtoParticipant : putLessonReq.getParticipants()) {
+                    if(lessonParticipant.getMemberId().equals(dtoParticipant.getMemberId())) {
+                        temp = Boolean.TRUE;
+                    }
+                }
+                if(!temp) {
+                    needUpdate = Boolean.FALSE;
+                    break;
+                }
+            }
+        }
+
+        // 숙제 비교
+        List<Homework> homeworkRowList = homeworkRepository.findAllByMemberIdAndLessonIdAndDeletedAtIsNull(member.getMemberId(), lessonId).orElse(null);
+        if(homeworkRowList != null && putLessonReq.getHomeworks() != null) {
+            if(homeworkRowList.size() == putLessonReq.getHomeworks().size()) {
+
+            }
+            if(homeworkRowList.size() != putLessonReq.getHomeworks().size()) {
+                needUpdate = Boolean.TRUE;
+            }
+        }
+
+
+        // 변경사항이 발생하면 숙제 업데이트 진행
+        if(needUpdate) {
+            homeworkRepository.deleteHomeworkByLessonId(lessonId);
+            List<Homework> homeworkInsertList = new ArrayList<>();
+            for (Participant participant : putLessonReq.getParticipants()) {
+                List<Homework> tempHomeworkList = putLessonReq.getHomeworks().stream()
+                        .map(homeworkReq -> HomeworkReq.staticToEntity(homeworkReq, participant.getMemberId(), lesson.getLessonId()))
+                        .collect(Collectors.toList());
+                homeworkInsertList.addAll(tempHomeworkList);
+            }
+            homeworkRepository.saveAll(homeworkInsertList);
+        }
 
         // Response
         return null;
